chore: move vue dependencies to plain dependencies
This allows in theory (tested) to also use this library with Vue 2 apps.
For those the bundle size will increase as Vue needs to bundled two
times, but if you use Vue 3 all works as intended (only bundled once so
small asset size).

This allows us to drop maintenance on v5 as soon as Nextcloud 31 gets
EOL.

Signed-off-by: Ferdinand Thiessen <opensource@fthiessen.de>