fix: ensure password is used for axios request
Before this the promise of `spawnDialog` was awaited so it never
resolved, because the dialog did not close until axios succeeded.
And axios only continued when the dialog was closed...

Instead add a module scope variable to keep track of an already open
password confirmation dialog and reuse it if the password is invalid.

Signed-off-by: Ferdinand Thiessen <opensource@fthiessen.de>